feat：Extract UIRefreshControl/UIButton test helper to a separate file
