+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -458,24 +458,3 @@ private extension FeedImageCell {
         feedImageView.image?.pngData()
     }
 }
-
-private extension UIButton {
-    func simulateTap() {
-        allTargets.forEach{ target in
-            actions(forTarget: target, forControlEvent: .touchUpInside)?.forEach{
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
-
-
-private extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        allTargets.forEach{ target in
-            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach{
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
