+++ EssentialFeed/EssentialFeediOSTests/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,18 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by nicaho on 2024/8/6.
+//
+
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        allTargets.forEach{ target in
+            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach{
+                (target as NSObject).perform(Selector($0))
+            }
+        }
+    }
+}
